added the feature engineering pipeline
