+++ conf/base/parameters_train_test_split_pipeline.yml
@@ -1,44 +1,3 @@
-train_test_split_pipeline.feature_engineering:
-  basic_features:
-    - dayofweek
-    - quarter
-    - month
-    - year
-    - dayofyear
-  column_names:
-    - total_consumption
-    - Global_intensity
-    - Sub_metering_3
-    - Sub_metering_1
-    - temp
-    - day_length
-    - tempmax
-    - feelslike
-    - feelslikemax
-    - feelslikemin
-    - tempmin
-  lags:
-    - 1
-    - 2
-    - 3
-    - 4
-    - 5
-    - 6
-    - 7
-    - 30
-    - 90
-    - 365
-  window_sizes:
-    - 2
-    - 3
-    - 4
-    - 5
-    - 6
-    - 7
-    - 30
-    - 90
-    - 365
-
 train_test_split_pipeline.feature_splitting:
   target:  total_consumption 
   threshold:  2010-05-17
