+++ src/energy_forcasting_model/pipelines/feature_engineering_pipeline/__init__.py
@@ -1,0 +1,10 @@
+"""
+This is a boilerplate pipeline 'feature_engineering_pipeline'
+generated using Kedro 0.19.3
+"""
+
+from .pipeline import create_pipeline
+
+__all__ = ["create_pipeline"]
+
+__version__ = "0.1"
