+++ src/tests/pipelines/feature_engineering_pipeline/test_pipeline.py
@@ -1,0 +1,9 @@
+"""
+This is a boilerplate test file for pipeline 'feature_engineering_pipeline'
+generated using Kedro 0.19.3.
+Please add your pipeline tests here.
+
+Kedro recommends using `pytest` framework, more info about it can be found
+in the official documentation:
+https://docs.pytest.org/en/latest/getting-started.html
+"""
